backend/requirements.txt
alembic~=1.9.2
aiohttp~=3.8.3
asyncpg~=0.25.0
asgi-lifespan~=2.0.0
black~=22.12.0
coverage~=7.1.0
dynaconf~=3.1.11
flake8~=6.0.0
fastapi~=0.92.0
google-auth~=2.16.2
httpx~=0.23.3
isort~=5.11.4
loguru~=0.5.3
mypy~=0.991
pydantic~=1.9.1
pyhumps~=3.8.0
pytz~=2022.7.1
psycopg2-binary~=2.9.5
pytest~=7.2.1
pytest-asyncio~=0.17.2
python-keycloak~=2.9.0
python-jose~=3.3.0
pytest-mock~=3.10.0
python-dotenv~=0.19.2
requests~=2.28.2
starlette~=0.25.0
SQLAlchemy~=1.4.46
uvicorn~=0.18.2
